update code example and update acc data
