+++ examples/language-modeling/run_autoround.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 set -x
 
-model="neural-chat-7b-v3-3"
+model="Intel/neural-chat-7b-v3"
 
 CUDA_VISIBLE_DEVICES=4 python3 main.py \
         --model_name /models/${model} \
@@ -13,5 +13,5 @@ CUDA_VISIBLE_DEVICES=4 python3 main.py \
         --deployment_device 'cpu' \
         --scale_dtype 'fp32' \
         --eval_bs 16 \
-        --output_dir "SAVE/PATH"
+        --output_dir "./tmp_signround"
